Move all benchmarking within deepsparse/benchmark/ (#333)
* Move all benchmarking within deepsparse/benchmark/

* Update benchmark_model

* Expose results at benchmark base

* isort

* Skip flake8
